fix: v3.53 经济系统平衡修复
P0: UPKEEP数组扩展至7个元素修复tier4-6维护费为0的崩坏
P1: 废品回收下限¥20->¥25避免升级住房死锁

+++ src/js/data/jobs.js
@@ -30,7 +30,8 @@ const STREET_JOBS = [
       agilityXp: 1,
     },
     payCalc(state) {
-      const base = Random.float(20, 55);
+      // v3.53 修复：下限从¥20提升到¥25，避免"升级住房→入不敷出"死锁
+      const base = Random.float(25, 55);
       const multi = 1 + state.skills.sales.level * 0.005;
       // v3.8 P1修复：zhouScrapBonus（老周好感奖励→废品+20%）
       const zhouBonus = state.flags && state.flags.zhouScrapBonus ? 1.2 : 1.0;
